save state

+++ src/app/components/Comment.tsx
@@ -2,10 +2,12 @@ import React, { useState, useEffect, ReactNode } from "react";
 import { iComment } from "../util/Interfaces";
 import dayjs from "dayjs";
 import Link from "next/link";
-import { ReplyIcon, PencilIcon, TrashIcon, SaveIcon, ChevronDownIcon, ChevronUpIcon } from 'lucide-react';
+import { ReplyIcon, SaveIcon, ChevronDownIcon, ChevronUpIcon } from 'lucide-react';
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import OptionsMenu from "./CommentOptionsMenu";
+import { useAuth } from "@clerk/nextjs";
 
 interface CommentProps {
   comment: iComment;
@@ -14,11 +16,13 @@ interface CommentProps {
   onDelete: (commentId: string) => Promise<void>;
   depth: number;
   children?: ReactNode;
+  clerkUserId: string;
 }
 
 const MAX_VISIBLE_DEPTH = 5;
 
-const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onEdit, onDelete, depth = 0 }) => {
+const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onEdit, onDelete, depth = 0, clerkUserId }) => {
+  const { userId } = useAuth();
   const [comment, setComment] = useState(initialComment);
   const [isEditing, setIsEditing] = useState(false);
   const [editedBody, setEditedBody] = useState(comment.body);
@@ -28,6 +32,10 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
   const [replies, setReplies] = useState<iComment[]>(comment.replies || []);
   const [showReplies, setShowReplies] = useState(depth < MAX_VISIBLE_DEPTH);
 
+  const isCommentOwner = clerkUserId === comment.user.clerkUserId
+    ;
+  const canReply = clerkUserId && clerkUserId !== comment.user.clerkUserId;
+
   useEffect(() => {
     setComment(initialComment);
     setEditedBody(initialComment.body);
@@ -35,7 +43,7 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
   }, [initialComment]);
 
   const handleReply = async () => {
-    if (comment.id) {
+    if (comment.id && canReply) {
       await onReply(comment.id, replyBody);
       const newReply: iComment = {
         id: Date.now().toString(),
@@ -44,7 +52,7 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
         createdDate: new Date(),
         review: comment.review,
         parentId: comment.id,
-        userId: comment.userId,
+        userId: userId as string,
         isDeleted: false,
         reviewId: comment.reviewId,
       };
@@ -56,13 +64,13 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
   };
 
   const handleEdit = () => {
-    if (comment.id) {
+    if (comment.id && isCommentOwner) {
       setIsEditing(true);
     }
   };
 
   const handleSave = async () => {
-    if (comment.id) {
+    if (comment.id && isCommentOwner) {
       await onEdit(comment.id, editedBody);
       const updatedComment = { ...comment, body: editedBody };
       setComment(updatedComment);
@@ -71,7 +79,7 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
   };
 
   const handleDelete = async () => {
-    if (comment.id && window.confirm("Are you sure you want to delete this comment?")) {
+    if (comment.id && isCommentOwner && window.confirm("Are you sure you want to delete this comment?")) {
       await onDelete(comment.id);
       const deletedComment = {
         ...comment,
@@ -93,6 +101,7 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
       <div className={`mt-2 ${depth >= MAX_VISIBLE_DEPTH - 1 ? 'border-l border-gray-200 pl-2' : 'ml-4'}`}>
         {replies.map((reply) => (
           <Comment
+            clerkUserId={clerkUserId}
             key={reply.id}
             comment={reply}
             onReply={onReply}
@@ -106,19 +115,22 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
   };
 
   return (
-    <div className={`w-full bg-white  border-b border-gray-100 p-3 ${depth > 0 ? 'ml-2' : ''}`}>
+    <div className={`w-full bg-white border-b border-gray-100 p-3 ${depth > 0 ? 'ml-2' : ''}`}>
       <div className="flex items-start space-x-2">
         <Avatar className="w-6 h-6">
           <AvatarImage src={comment.user?.avatar || "/default-avatar.png"} alt={`${comment.user?.firstName} ${comment.user?.lastName}`} />
           <AvatarFallback>{comment.user?.firstName?.charAt(0)}{comment.user?.lastName?.charAt(0)}</AvatarFallback>
         </Avatar>
         <div className="flex-1 min-w-0">
-          <div className="flex items-center text-xs text-gray-500 mb-1">
-            <Link href={`/userprofile/${comment.user?.id}`} className="font-medium text-blue-600 hover:underline">
-              @{comment.user?.userName}
-            </Link>
-            <span className="mx-1">•</span>
-            <span>{dayjs(comment.createdDate).format("MMM D, YYYY")}</span>
+          <div className="flex items-center justify-between">
+            <div className="flex items-center text-xs text-gray-500 mb-1">
+              <Link href={`/userprofile/${comment.user?.id}`} className="font-medium text-blue-600 hover:underline">
+                @{comment.user?.userName}
+              </Link>
+              <span className="mx-1">•</span>
+              <span>{dayjs(comment.createdDate).format("MMM D, YYYY")}</span>
+            </div>
+            {!comment.isDeleted && isCommentOwner && <OptionsMenu onEdit={handleEdit} onDelete={handleDelete} setIsEditing={setIsEditing} />}
           </div>
           <div className="text-sm text-gray-700">
             {isEditing ? (
@@ -143,27 +155,20 @@ const Comment: React.FC<CommentProps> = ({ comment: initialComment, onReply, onE
               </>
             )}
           </div>
-          {!comment.isDeleted && (
+          {!comment.isDeleted && !isEditing && (
             <div className="mt-2 flex flex-wrap gap-2">
-              <Button variant="ghost" size="sm" className="text-xs px-2 py-0 h-6" onClick={() => setIsReplying(!isReplying)}>
-                <ReplyIcon className="w-3 h-3 mr-1" />
-                Reply
-              </Button>
-              {isEditing ? (
+              {canReply && (
+                <Button variant="ghost" size="sm" className="text-xs px-2 py-0 h-6" onClick={() => setIsReplying(!isReplying)}>
+                  <ReplyIcon className="w-3 h-3 mr-1" />
+                  Reply
+                </Button>
+              )}
+              {isEditing && isCommentOwner && (
                 <Button variant="ghost" size="sm" className="text-xs px-2 py-0 h-6" onClick={handleSave}>
                   <SaveIcon className="w-3 h-3 mr-1" />
                   Save
                 </Button>
-              ) : (
-                <Button variant="ghost" size="sm" className="text-xs px-2 py-0 h-6" onClick={handleEdit}>
-                  <PencilIcon className="w-3 h-3 mr-1" />
-                  Edit
-                </Button>
               )}
-              <Button variant="ghost" size="sm" className="text-xs px-2 py-0 h-6 text-red-500" onClick={handleDelete}>
-                <TrashIcon className="w-3 h-3 mr-1" />
-                Delete
-              </Button>
             </div>
           )}
           {isReplying && (
